Mapped away check command
Took 2 minutes
